# Replace `fork-ts-checker-webpack-plugin` with faster alternative