--- a/packages/next/lib/typescript/TypeScriptCompileError.ts
+++ b/packages/next/lib/typescript/TypeScriptCompileError.ts
@@ -1,0 +1,1 @@
+export class TypeScriptCompileError extends Error {}