--- a/packages/next/lib/verifyTypeScriptSetup.ts
+++ b/packages/next/lib/verifyTypeScriptSetup.ts
@@ -1,3 +1,4 @@
+import chalk from 'next/dist/compiled/chalk'
 import path from 'path'
 import { FatalTypeScriptError } from './typescript/FatalTypeScriptError'
 import { getTypeScriptIntent } from './typescript/getTypeScriptIntent'
@@ -5,20 +6,23 @@ import {
   hasNecessaryDependencies,
   NecessaryDependencies,
 } from './typescript/hasNecessaryDependencies'
+import { runTypeCheck, TypeCheckResult } from './typescript/runTypeCheck'
+import { TypeScriptCompileError } from './typescript/TypeScriptCompileError'
 import { writeAppTypeDeclarations } from './typescript/writeAppTypeDeclarations'
 import { writeConfigurationDefaults } from './typescript/writeConfigurationDefaults'
 
 export async function verifyTypeScriptSetup(
   dir: string,
-  pagesDir: string
-): Promise<void> {
+  pagesDir: string,
+  typeCheckPreflight: boolean
+): Promise<TypeCheckResult | boolean> {
   const tsConfigPath = path.join(dir, 'tsconfig.json')
 
   try {
     // Check if the project uses TypeScript:
     const intent = await getTypeScriptIntent(dir, pagesDir)
     if (!intent) {
-      return
+      return false
     }
     const firstTimeSetup = intent.firstTimeSetup
 
@@ -35,9 +39,19 @@ export async function verifyTypeScriptSetup(
     // Write out the necessary `next-env.d.ts` file to correctly register
     // Next.js' types:
     await writeAppTypeDeclarations(dir)
+
+    if (typeCheckPreflight) {
+      // Verify the project passes type-checking before we go to webpack phase:
+      return await runTypeCheck(ts, dir, tsConfigPath)
+    }
+    return true
   } catch (err) {
-    // This is a special error that should not show its stack trace:
-    if (err instanceof FatalTypeScriptError) {
+    // These are special errors that should not show a stack trace:
+    if (err instanceof TypeScriptCompileError) {
+      console.error(chalk.red('Failed to compile.\n'))
+      console.error(err.message)
+      process.exit(1)
+    } else if (err instanceof FatalTypeScriptError) {
       console.error(err.message)
       process.exit(1)
     }