--- a/packages/next/package.json
+++ b/packages/next/package.json
@@ -86,7 +86,6 @@
     "chokidar": "2.1.8",
     "css-loader": "3.5.3",
     "find-cache-dir": "3.3.1",
-    "fork-ts-checker-webpack-plugin": "3.1.1",
     "jest-worker": "24.9.0",
     "loader-utils": "2.0.0",
     "mini-css-extract-plugin": "0.8.0",