@johnnyreilly 2020-05-29
You will no doubt will (and should!) pick the approach that works best for you. Couple of things to note in case there are properties of the `fork-ts-checker-webpack-plugin` that are directly useful to the project:

- it looks like you're using an old version (V3?) of the checker - I wonder how v4 compares?
- related to that `piotr-oles` has been doing some incredible work for v5 https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/pull/404 - even though it's technically alpha it's pretty well advanced and I'd recommend taking it for a spin!

Of course, please make the calls that work for your project, I just wanted to let you know some possibilities 🤗