@johnnyreilly 2020-05-29
Thanks for the detailed response @Timer! It does sound like you're planning some pretty hefty (and cool sounding!) architectural changes. Super interesting that you don't run type checking in development - that's where I particularly appreciate it!

> We want to run type checking before the webpack build, and skip it [webpack] if checks fail. This gives a faster turnaround time IMO.

This is a really interesting idea. I'm very curious as to how it works out.

> We firmly believe in keeping the webpack build's responsibilities as small as possible. We're going to start moving as many things as we can into side-cars, completely off the Node.js event loop (this PR is the first step).

Super interesting - as the name suggests, the initial motivation for `fork-ts-checker-webpack-plugin` was exactly that: "Webpack plugin that runs TypeScript type checker on a separate process."

Sounds like you're going all in on this. Fascinating.

If you do get a chance to compare v4 and v5 alpha performance we'd love to get an idea of the numbers.

Thanks again for sharing the ideas here - really cool!