@johnnyreilly 2020-05-29
Thanks @timer - that's tremendously helpful! @piotr-oles it looks like there's a performance degradation with v5 - see the numbers from @timer above.