@johnnyreilly 2020-05-29
> We're also exploring ESLint in the future too, so this all parlays.

Did you use the ESLint integration in fork-ts-checker-webpack-plugin BTW? My own use case is running `fork-ts-checker-webpack-plugin` to surface up both type checking errors from TypeScript *and* lints from ESLint.