@johnnyreilly 2020-05-29
> The DX we're aiming for (in dev) is that TS/ESLint errors show up in a toast though, and not a full overlay ("compile error").

Yeah that was exactly why I wrote https://github.com/johnnyreilly/fork-ts-checker-notifier-webpack-plugin

Something very lovely about having errors as toasts outside of the application. I've never really enjoyed the full overlay approach; I like to have the application running even if it does contain errors. I find value in being able to debug a broken application - particularly if it's a type system error which may not massively impair functionality.