@johnnyreilly 2020-05-29
Incidentally, I've an idea that `isolatedModules: true` may hurt your performance and `skipLibCheck: true` could improve your performance (regardless of context).

Incidentally, `skipLibCheck: true` is going to become the default as of TypeScript 3.9