@piotr-oles 2020-05-29
@Timer Thanks for the benchmark, I will try to investigate why there is such a big difference between `tsc` and `fork-ts-checker-webpack-plugin` since we use the same API 😃