@timneutkens 2020-05-29
> skipLibCheck: true could improve your performance (regardless of context).

It's already true in the benchmarks @Timer did (so didn't affect the slowdown)