@piotr-oles 2020-05-29
@Timer Could you tell me how can I check timings on my local machine? I cloned the repository and checkout master branch but I'm not sure which command I should run to use fork-ts-checker-webpack-plugin :)