@piotr-oles 2020-05-29
@Timer Ok, I will try on different repository :) Could you tell me which version of the typescript did you use for tests?