@timneutkens 2020-05-29
```
    "typescript": "3.7.2"
```