@piotr-oles 2020-05-29
Thanks :) If someone has a project that can share and that is significantly slower with `fork-ts-checker-webpack-plugin` than with `tsc`, it would be very helpful for testing 😃