@piotr-oles 2020-06-03
I did some research regarding plugin performance and these are outcomes:
 * the `fork-ts-checker-webpack-plugin` is slower than `tsc` partially because it runs in parallel with the `webpack`. I'm not sure about the cause of it but I would bet on the file system. It was about 20% slower in my case.
 * there was also some difference even if the plugin ran without webpack compilation in parallel. I didn't found the cause (was about 15% slower).
 * I improved the initial compilation time by using `incremental: true` by default. The time dropped from ~100 s to ~30 s (https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/pull/435)
 * I fixed watcher implementation for incremental compilation. Thanks to that the incremental time dropped from ~15 s to ~2 s (https://github.com/TypeStrong/fork-ts-checker-webpack-plugin/pull/435)